requirements.txt
streamlit
google-generativeai
beautifulsoup4
requests
PyMuPDF
chromadb
langchain
sentence-transformers
python-dotenv